invest_simulator_backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
supabase==1.0.4
yfinance==0.2.28
pandas==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
requests==2.31.0
APScheduler==3.10.4
twilio==8.2.2